Add files via upload
A BMI (Body Mass Index) generator is a tool or software that calculates an individual's BMI based on their height and weight. The BMI is a numerical value that indicates whether a person is underweight, normal weight, overweight, or obese. It's a quick and easy way to assess body fatness and potential health risks.